settings from args or json:
Adding a function checking whether a setting is defined in json and if
it has been overridden by arguments.

The settings initialized with this function have to be at least defined
in the json file.

The limitation here is that these settings are mandatory even if they
are not use in practice. For example, defining package_final_phase is
not necessary in install mode.